+++ GUI/widgets/timer.py
@@ -7,8 +7,8 @@ import sys
 class Timer(QWidget):
     def __init__(self, *args, parent=None):
         super(Timer, self).__init__(parent)
-        self.width = args[0] / 10
-        self.height = args[0] / 30
+        self.width = int(args[0] / 10)
+        self.height = int(args[0] / 30)
         self.initUI()
 
     def initUI(self):
@@ -19,7 +19,7 @@ class Timer(QWidget):
         # self.label.setFont(QFont('AnyStyle', 14))
         self.label.setStyleSheet(
             "font-family: Helvetica; font-size: 14px; background-color: #2B26c1; color: black")
-        self.label.resize(self.width, self.height)
+        self.label.resize(int(self.width), int(self.height))
         self.timer = QTimer(self)
         self.timer.timeout.connect(self.showTime)
         self.timer.start(100)
@@ -31,4 +31,4 @@ class Timer(QWidget):
         self.label.setText(text)
 
     def sizeHint(self):
-        return QSize(self.width, self.height)
+        return QSize(int(self.width), int(self.height))
